+++ packages/vsce/src/resources/actions/LibraryActions.ts
@@ -41,13 +41,6 @@ const actions: ResourceActionOptions<ResourceTypes.CICSLibrary>[] = [
     action: "cics-extension-for-zowe.copyResourceName",
     refreshResourceInspector: false,
   },
-  {
-    id: "CICS.CICSLibrary.VIEW_IN_TABLE",
-    name: l10n.t("View in Table"),
-    resourceType: ResourceTypes.CICSLibrary,
-    action: "cics-extension-for-zowe.viewInTable",
-    refreshResourceInspector: false,
-  },
 ];
 
 export function getLibraryActions(): ResourceAction<ResourceTypes.CICSLibrary>[] {